Changes STSmobile and STSX
	clearDatabases create pre-import module indexes to tables for instances
expected in each window
	fs.js interface fabsuite check
	fs status check
	fs warnings titles to i18n content, not error number
	fs UN UserName check. NOOP
	relations add assoc to plant, i18n, idfile->current_load/ receive_load
/ship_load, job_ship_to, i18n_language
	valuelists unprotected_groups, jobs_assoc_thru_relation
	globals.js mob.job.rf cache setting for id whether it is connected
through 3rd party
	mob.load cache for total, shipped, loadFs
	mob.userAgent for current device identifier from web browser when in
STSmobile web client
	session.adminUser to determine table access
	session.corpUser that user is corp level
	session.dualEntry for IOS specific entry issues on single entry losing
focus, so forces double entry
	session.interim interim ship or not
	session.parentAssocId is company record, not assoc/division
	session.functionName determine in which function the user is operating
	rfViews.{screen}.genericin2 added with genericin to get uninterrupted
entry on IOS, using global setting for visibility
	rfViews'Transactions' locationPieces and locationWeight moved after ID
	i18n localeSelected for i18n table entries
	functionKey handling on F2
	F8 was sticky, reset upon leaving screen
	location changes were getting lost on trailer input
	cached duplicate workers check
	barcode length preferences and return correct barcode for new
preference
	checkLegs on route condensed code
	rfF3 ids not shipped view
	rfF4 pcmks not shipped view
	rfF5 load status view
	rfF6 print packing list
	rfF7 trailer info view
	bundle vs barcode vs part ID with variable length barcodes
	rfErrorShow dual entry handling
	rfActivity was not i18n and needed expansion to cover more screens
	rfSaveScanTransaction isManHours  to indicate data saved as man hours
	F1 show help wasn't deactivated upon close of screen
	Choosing Exit on browser will kill off the client internally,
refreshing the page to unavailable
	Internal changes to table column names for consistency
	Save additional values into loads
	Loads info reporting more than counted
	getSpecsLoad cleaner calculation reporting between load types
	jobnumber caching and code style changes ie predefine rec and cycle
through records
	entryType on generic input is set as a prefix on screen to view last
command interpretation
	re-use of some entry types per screen, running out of characters
	click STSmobile plant button will show current browser information
	silentPrint for STSmobile
	renderPrinterButtons
	silentPrintToServer STSmobile
	renderPackingButtons
	rfFinalShip
	jobsts.jobs.js importjob.saveNotesInto / savePhasePcColor /
notesContainCamber / importArea / importRouting / keepMinors /
userImportRouting / jobMetric
	browseLoads changed to collection and show table
	inactive tabs on browseLoads
	kiss import location not updating
	kiss import either use route codes in import or selected on screen
	label quantity calculations on import as per label and total quantity
	non-odometer countup for barcodes [A-Z]*[0-9]* instead of each column
countup from [0-9A-Z]
	record selection for browseLoads locking to singular transaction view,
was a problem with form instances
	protection for tables from corp vs plant users
	cleaned import logic KISS
	printer templates changes to barcode printer
	import_table form changes
	parent barcode for a particular barcode wasn't tracked
	clear intermediate import_table on KISS before prepping the next for
import
	repeatable import settings stored to preferences
	add edit date to secGroupBase
	permissions update showing variable checked items
	parent assoc relation added
	servoySecurity globals.js client userAgent for identifying
device/browser for IOS web page interaction
	changes to new install tables cleared and initial account creation
	kill off clients that are idle when there are no more licenses
available as assigned
	upon login attempt set global userAgent variable  for browser info
	set STSmobile function menu and screen  menu to show partial height on
Android/iPhone
	employee addresss not setting correctly
	verify input fields as to minimum required to Save action to
apply/appear
	Cancel Edit moved to base screen for editing addresses
	Easily copy user group in permissions
	barcode label form setting label formats
	bartender print workaround from seagull caused by COM bridge by Servoy
	cost of work category screen changes and adjustment 
	cow form clearing
	cow not showing all missing codes
	division_user_detail input verification
	division_user input  verification
	other forms input verification
	end_conditions close  window track
	i18n_edit screen
	jobs screen changes
	kiss import screen changes and locks, minor marks inclusion
	piecemark entry screen
	preferences paths integration with servoyforge default directories
	integrate servoyforge plugin velocity reports
	improved busy progress popup to long wait areas
	route code editing
	add columns and functions to view licensing screen
	STSmobile viewport scaling IOS Android

+++ servoySecurity/forms/secUserBase.js
@@ -18,7 +18,10 @@ var userGroups = null;
  * @properties={typeid:35,uuid:"3AD72C25-2E35-4E73-B980-02CDFAD81557"}
  */
 var userPassword = null;
-
+/**
+ * @properties={typeid:35,uuid:"EEABB406-2720-4A68-AD2F-2C757651BE8C",variableType:-4}
+ */
+var changeTime = null;
 /**
  * Handle changed data for the userGroups variable
  * Which stores a carriage return-delimted list of linked Group ids
@@ -35,6 +38,9 @@ var userPassword = null;
  * @AllowToRunInFind
  */
 function onDataChangeUserGroups(oldIDs, newIDs, event) {
+	var changeDelta = new Date().getTime()-changeTime;
+	if (changeDelta < 300){return true}
+	
 																			//	Variable Declarations
 	var fs = users_to_user_groups.duplicateFoundSet();						//	The related foundset 
 	var idColumnName = 'group_uuid';											//	The name of the fk id column to set
@@ -50,8 +56,10 @@ function onDataChangeUserGroups(oldIDs, newIDs, event) {
 	if(newIDs.length > oldIDs.length){										//	check difference: id was ADDED
 		id = newIDs[newIDs.length - 1];									//	determine the id, starting with the last element in the array
 		for(i in oldIDs){													//	iterate by the old ids
-			if(oldIDs[i] != newIDs[i])										//	until the corresponding new id doesn't match
+			if(oldIDs[i] != newIDs[i]){										//	until the corresponding new id doesn't match
 				id = newIDs[i];												//	which gives the id that was added
+				break;
+			}
 		}
 		if(fs.newRecord()){													//	create the related record
 			fs[idColumnName] = id;											//	set the foreign key column value
@@ -59,19 +67,35 @@ function onDataChangeUserGroups(oldIDs, newIDs, event) {
 			fs.edit_date = new Date();
 			databaseManager.saveData(fs.getSelectedRecord());				//	save data
 		}
-	} else {																//	id was REMOVED	
-		id = oldIDs[oldIDs.length - 1];									//	determine the id, starting with the last element in the array
-		for(i in newIDs){													//	iterate by the new ids
-			if(oldIDs[i] != newIDs[i])										//	until the corresponding old id doesn't match
-				id = oldIDs[i];												//	which gives the id that was removed
+		for (i in oldIDs){
+			if (oldIDs[i] == id){continue}
+			if(fs.find()){														//	search the foundset
+				fs[idColumnName] = oldIDs[i];											//	by the fk column
+				if(fs.search())																					
+					fs.deleteRecord();											//	delete record
+			}
 		}
-		if(fs.find()){														//	search the foundset
-			fs[idColumnName] = id;											//	by the fk column
-			if(fs.search())																					
-				fs.deleteRecord();											//	delete record
+	} else {//delete the key		
+		id = oldIDs[oldIDs.length - 1];									//	determine the id, starting with the last element in the array
+		//for (i = 0;i < newIDs.length;i++){
+		for(i in newIDs){													//	iterate by the new ids
+			if(oldIDs[i] != newIDs[i]){										//	until the corresponding old id doesn't match
+				id = oldIDs[i];												//	which gives the id that was removed
+				break;
+			}
+		}
+		for (i in oldIDs){
+			//if (oldIDs[i] == id){continue}
+			if(fs.find()){														//	search the foundset
+				fs[idColumnName] = id;											//	by the fk column
+				if(fs.search())																					
+					fs.deleteRecord();											//	delete record
+			}
 		}
 	}
 	updateUI(event);																//	update user interface
+	changeTime = new Date().getTime();
+	if (forms['division_user_details']){forms['division_user_details'].verifyNewUserInput(event);}
 	return true;															//	allow data change
 }
 
@@ -170,8 +194,14 @@ function validateUserName(event){
 		errorMessage = i18n.getI18NMessage('sts.txt.provide.user.name');
 		return false;														//	failed validation
 	}
+	databaseManager.removeTableFilterParam('stsservoy','filterAssocUSERS');//disable USER Filters for user existence check
 	var id = globals.secGetUserID(user_name,tenant_uuid);						//	Check user ID
+	if (!globals.session.corpUser){
+		databaseManager.addTableFilterParam('stsservoy','users','association_uuid','=',globals.session.associationId,'filterAssocUSERS');//Re-enable user assoc table filters
+	}
+
 	if(id && id != user_uuid){												//	It should be unique
+		plugins.dialogs.showErrorDialog('1260',i18n.getI18NMessage('1260'));//User Name Already In Use In A Division.
 		errorMessage = i18n.getI18NMessage('sts.txt.user.name.already.in.use');
 		/** @type {QBSelect<db:/stsservoy/users>} */
 		return false;														//	failed validation
